requirements.txt
langchain
langchain-openai
langchain-community
pypdf
pydantic
streamlit
python-dotenv
pandas